+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - Add basic examples about how to use the module as a library
 - Relax dependencies to be able to use a more recent version of Pandas
 - Be graceful about malformed lines within ingress CSV
+- Work around deprecated Pandas feature regarding index vs. column ambiguity
 
 2020-01-07 0.9.1
 ================